requirements.txt
pymupdf==1.23.8
pytesseract==0.3.10
Pillow==10.2.0
